Structuration Restfull & delete of filter for now
